Rock, Paper, Scissors, Lizard, Spock!!!
I changed the win-and-lose check so I wouldn't go insane and so it is easier to read. Also, I didn't know you could do it like this in the first place. Now both the player and computer have more options. I also changed the rules explanation to keep the joke going.